Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src-attr-elem/style-src-elem-allowed-src-blocked-expected.txt


FAIL Inline style should be applied assert_equals: expected 1 but got 0
